+++ NEWS.md
@@ -1,6 +1,7 @@
 1.6.1 - 2013-MM-DD
 ------------------
 * Still set a meaningful X-Cache-Channel with cache_policy=persist (#105)
+* Fix wrong projection in KML exports for manually altered tables (#116)
 
 1.6.0 - 2013-10-02
 ------------------