+++ NEWS.md
@@ -1,4 +1,4 @@
-1.6.1 - 2013-MM-DD
+1.6.1 - 2013-11-05
 ------------------
 * Still set a meaningful X-Cache-Channel with cache_policy=persist (#105)
 * Fix wrong projection in KML exports for manually altered tables (#116)